+++ srcs/cpp/include/kungfu/nccl/common.hpp
@@ -1,0 +1,26 @@
+#pragma once
+#include <map>
+#include <string>
+
+#include <kungfu.h>
+
+enum KungFu_NCCLScope {
+    KungFu_NCCL_GLOBAL,
+    KungFu_NCCL_LOCAL,
+};
+
+namespace kungfu
+{
+// Workspace holds metadata for AllReduce, Reduce, Broadcast
+struct Workspace {
+    const void *sendbuf;
+    void *recvbuf;
+    const int count;
+    const KungFu_Datatype dtype;
+};
+
+void CrossAllReduceGpu(const Workspace &w, KungFu_Op op,
+                       const std::string &name, DoneCallback done);
+
+extern const std::map<std::string, KungFu_NCCLScope> _nccl_scopes;
+}  // namespace kungfu
