+++ srcs/cpp/include/kungfu/nccl/controller.hpp
@@ -1,0 +1,24 @@
+#pragma once
+#include <kungfu/nccl/common.hpp>
+#include <kungfu/nccl/gpu_collective.hpp>
+
+namespace kungfu
+{
+class NCCLController
+{
+    KungFu_NCCLScope scope_;
+    std::unique_ptr<gpu_collective> gpu_collective_;
+
+  public:
+    NCCLController(const KungFu_NCCLScope scope);
+
+    void InitOnce();
+
+    int Reduce(const Workspace &w, KungFu_Op op, DoneCallback done);
+
+    int Broadcast(const Workspace &w, DoneCallback done);
+
+    int AllReduce(const Workspace &w, KungFu_Op op, DoneCallback done);
+};
+
+}  // namespace kungfu
